pyproject.toml
[project]
name = "big-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "big-vision",
    "chex>=0.1.86",
    "distrax>=0.1.5",
    "dlimp@git+https://github.com/kvablack/dlimp.git",
    "einops>=0.8.0",
    "flax>=0.9.0",
    "ipykernel",
    "jax>=0.4.34",
    "matplotlib>=3.9.2",
    "ml-collections>=0.1.1",
    "numpy<2.0.0",
    "octo@git+https://github.com/octo-models/octo.git",
    "optax>=0.2.3",
    "orbax-checkpoint>=0.7.0",
    "orbax>=0.1.9",
    "overrides>=7.7.0",
    "pip",
    "scalax>=0.2.4",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "sentencepiece>=0.2.0",
    "tfds-nightly>=4.9.0",
    "tensorflow-text-nightly>=2.15.0",
    "tf-nightly>=2.15.0",
    "wandb>=0.18.3",
    "protobuf>=3.20",
    "huggingface-hub>=0.27.0",
    "transformers>=4.47.1",
    "tensorflow-metadata==1.16.1",
]

[project.optional-dependencies]
tpu = [
    "jax[tpu]>=0.4.34",
    "libtpu-nightly",
]
gpu = [
    "jax[cuda12]>=0.4.34"
]

[tool.uv]
find-links = ["https://storage.googleapis.com/jax-releases/libtpu_releases.html"]
prerelease = "allow"
conflicts = [
    [
        { extra = "tpu" },
        { extra = "gpu" },
    ],
]
override-dependencies = [
    # Always use tf-nightly and tfds-nightly instead of tensorflow and tensorflow_datasets
    "tensorflow ; sys_platform == 'never'",
    "tensorflow_datasets ; sys_platform == 'never'",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
big-vision = { workspace = true }
